refactor(server): split registry model contract
