+++ crates/temper-server/src/registry/mod.rs
@@ -5,6 +5,9 @@
 //! in `ServerState`, enabling multi-tenant deployments where each tenant has
 //! its own entity types and specs.
 
+mod model_contract;
+#[cfg(test)]
+mod model_contract_tests;
 mod relations;
 pub mod types;
 #[cfg(test)]
@@ -28,6 +31,7 @@ use crate::trigger::types::ReactionRule;
 
 pub use types::*;
 
+use model_contract::model_safety_contract_changed;
 use relations::{build_relation_graph, build_webhook_routes, synthesize_action_trigger_reaction};
 
 fn merge_reaction_rules(
@@ -43,117 +47,6 @@ fn merge_reaction_rules(
         merged.insert(rule.name.clone(), rule);
     }
     merged.into_values().collect()
-}
-
-fn model_safety_contract_changed(
-    current: &automaton::Automaton,
-    incoming: &automaton::Automaton,
-) -> bool {
-    let current_runtime = automaton::compile_runtime_invariants(current);
-    let incoming_runtime = automaton::compile_runtime_invariants(incoming);
-    let current_has_model_contract = current.invariants.len() > current_runtime.len();
-    let incoming_has_model_contract = incoming.invariants.len() > incoming_runtime.len();
-    if !current_has_model_contract && !incoming_has_model_contract {
-        return false;
-    }
-
-    let invariant_signatures = |spec: &automaton::Automaton| {
-        spec.invariants
-            .iter()
-            .map(|invariant| {
-                let mut when = invariant.when.clone();
-                when.sort();
-                (invariant.name.clone(), when, invariant.assert.clone())
-            })
-            .collect::<std::collections::BTreeSet<_>>()
-    };
-    if invariant_signatures(current) != invariant_signatures(incoming) {
-        return true;
-    }
-
-    model_reachability_semantics(current) != model_reachability_semantics(incoming)
-}
-
-#[derive(PartialEq)]
-struct ModelReachabilitySemantics {
-    status_values: Vec<String>,
-    initial_status: String,
-    state_initials: BTreeMap<String, (String, serde_json::Value)>,
-    actions: Vec<ModelActionSemantics>,
-}
-
-#[derive(PartialEq)]
-struct ModelActionSemantics {
-    name: String,
-    from_states: Vec<String>,
-    to_state: Option<String>,
-    guard: automaton::ResolvedGuard,
-    effects: Vec<automaton::ResolvedEffect>,
-}
-
-fn model_reachability_semantics(spec: &automaton::Automaton) -> ModelReachabilitySemantics {
-    let states = spec
-        .state
-        .iter()
-        .filter(|state| matches!(state.var_type.as_str(), "bool" | "counter" | "list" | "set"))
-        .map(|state| {
-            (
-                state.name.clone(),
-                (
-                    state.var_type.clone(),
-                    automaton::parse_var_initial_json(&state.var_type, &state.initial),
-                ),
-            )
-        })
-        .collect::<BTreeMap<_, _>>();
-    let actions = automaton::translate_actions(spec)
-        .into_iter()
-        .map(|action| {
-            let effects = action
-                .effects
-                .into_iter()
-                .filter(automaton::ResolvedEffect::is_verifiable)
-                .collect::<Vec<_>>();
-            ModelActionSemantics {
-                name: action.name,
-                from_states: action.from_states,
-                to_state: action.to_state,
-                guard: normalized_model_guard(action.guard),
-                effects,
-            }
-        })
-        .collect::<Vec<_>>();
-
-    let mut status_values = spec.automaton.states.clone();
-    status_values.sort();
-    ModelReachabilitySemantics {
-        status_values,
-        initial_status: spec.automaton.initial.clone(),
-        state_initials: states,
-        actions,
-    }
-}
-
-fn normalized_model_guard(guard: automaton::ResolvedGuard) -> automaton::ResolvedGuard {
-    match guard {
-        automaton::ResolvedGuard::CrossEntityState {
-            entity_type,
-            entity_id_source,
-            required_status,
-            forbidden_status,
-            required: _,
-        } => automaton::ResolvedGuard::CrossEntityState {
-            entity_type,
-            entity_id_source,
-            required_status,
-            forbidden_status,
-            required: false,
-        },
-        automaton::ResolvedGuard::And(guards) => {
-            automaton::ResolvedGuard::And(guards.into_iter().map(normalized_model_guard).collect())
-        }
-        other => other,
-    }
 }
 
 /// Multi-tenant specification registry.
@@ -1065,344 +958,6 @@ assert = "title != ''"
                 .ioa_source,
             original_source,
             "rejected activation must leave the prior registry untouched"
-        );
-    }
-
-    #[test]
-    fn hot_swap_rejects_guard_only_model_contract_before_mutating_registry() {
-        const OLD_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[state]]
-name = "unlock"
-type = "bool"
-initial = "false"
-
-[[action]]
-name = "BreakSafety"
-kind = "input"
-from = ["Active"]
-to = "Active"
-guard = [{ type = "is_true", var = "unlock" }]
-effect = [{ type = "set_bool", var = "safe", value = "false" }]
-"#;
-        const VERIFIED_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[state]]
-name = "unlock"
-type = "bool"
-initial = "false"
-
-[[action]]
-name = "BreakSafety"
-kind = "input"
-from = ["Active"]
-to = "Active"
-guard = [{ type = "is_true", var = "unlock" }]
-effect = [{ type = "set_bool", var = "safe", value = "false" }]
-
-[[invariant]]
-name = "AlwaysSafe"
-when = ["Active"]
-assert = "safe"
-"#;
-
-        let mut registry = SpecRegistry::new();
-        let (csdl, xml) = minimal_csdl();
-        registry.register_tenant("alpha", csdl, xml, &[("Order", OLD_IOA)]);
-        let tenant = TenantId::new("alpha");
-        let original_spec = registry.get_spec(&tenant, "Order").unwrap();
-        let original_lock = original_spec.swap_controller().current();
-        let original_table = original_spec.table();
-        assert!(
-            original_table.model_protected_state_vars.is_empty(),
-            "the old contract permits durable caller-authored boolean state"
-        );
-        let original_table_json = serde_json::to_value(&*original_table).unwrap();
-        let original = registry.get_tenant(&tenant).unwrap();
-        let original_source = original.entities["Order"].ioa_source.clone();
-        let original_csdl = Arc::clone(&original.csdl);
-        let original_csdl_xml = Arc::clone(&original.csdl_xml);
-
-        let (replacement_csdl, replacement_xml) = minimal_csdl();
-        let error = registry
-            .try_register_tenant(
-                "alpha",
-                replacement_csdl,
-                replacement_xml,
-                &[("Order", VERIFIED_IOA)],
-            )
-            .expect_err("a new reachability contract requires durable-state migration");
-
-        assert!(matches!(
-            error,
-            RegistryError::RuntimeInvariantMigrationRequired { .. }
-        ));
-        let current = registry.get_tenant(&tenant).unwrap();
-        assert_eq!(current.entities["Order"].ioa_source, original_source);
-        assert!(Arc::ptr_eq(&current.csdl, &original_csdl));
-        assert!(Arc::ptr_eq(&current.csdl_xml, &original_csdl_xml));
-        let current_spec = registry.get_spec(&tenant, "Order").unwrap();
-        let current_lock = current_spec.swap_controller().current();
-        assert!(Arc::ptr_eq(&current_lock, &original_lock));
-        let current_table = current_spec.table();
-        assert_eq!(
-            serde_json::to_value(&*current_table).unwrap(),
-            original_table_json,
-            "rejection must leave the prior live transition table untouched"
-        );
-    }
-
-    #[test]
-    fn hot_swap_allows_metadata_only_change_under_same_model_contract() {
-        const ORIGINAL_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[invariant]]
-name = "AlwaysSafe"
-when = ["Active"]
-assert = "safe"
-"#;
-        const METADATA_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-overflow_inline_max_bytes = 4096
-overflow_ttl_seconds = 60
-
-[[invariant]]
-name = "AlwaysSafe"
-when = ["Active"]
-assert = "safe"
-"#;
-
-        let mut registry = SpecRegistry::new();
-        let (csdl, xml) = minimal_csdl();
-        registry.register_tenant("alpha", csdl, xml, &[("Order", ORIGINAL_IOA)]);
-        let original_lock = registry
-            .get_spec(&TenantId::new("alpha"), "Order")
-            .unwrap()
-            .swap_controller()
-            .current();
-
-        let (replacement_csdl, replacement_xml) = minimal_csdl();
-        registry
-            .try_register_tenant(
-                "alpha",
-                replacement_csdl,
-                replacement_xml,
-                &[("Order", METADATA_IOA)],
-            )
-            .expect("runtime-only metadata must remain hot-swappable");
-
-        let current = registry.get_spec(&TenantId::new("alpha"), "Order").unwrap();
-        assert!(Arc::ptr_eq(
-            &current.swap_controller().current(),
-            &original_lock
-        ));
-        let metadata = &current.table().state_var_metadata["safe"];
-        assert_eq!(metadata.overflow_inline_max_bytes, Some(4096));
-        assert_eq!(metadata.overflow_ttl_seconds, Some(60));
-    }
-
-    #[test]
-    fn hot_swap_rejects_reachability_change_under_same_model_invariant() {
-        const GUARDED_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[state]]
-name = "unlock"
-type = "bool"
-initial = "false"
-
-[[action]]
-name = "ChangeReachability"
-kind = "input"
-from = ["Active"]
-to = "Active"
-guard = [{ type = "is_true", var = "unlock" }]
-effect = [{ type = "set_bool", var = "unlock", value = "false" }]
-
-[[invariant]]
-name = "AlwaysSafe"
-when = ["Active"]
-assert = "safe"
-"#;
-        const REACHABLE_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Active"]
-initial = "Active"
-allow_indefinite_states = ["Active"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[state]]
-name = "unlock"
-type = "bool"
-initial = "false"
-
-[[action]]
-name = "ChangeReachability"
-kind = "input"
-from = ["Active"]
-to = "Active"
-guard = [{ type = "is_false", var = "unlock" }]
-effect = [{ type = "set_bool", var = "unlock", value = "true" }]
-
-[[invariant]]
-name = "AlwaysSafe"
-when = ["Active"]
-assert = "safe"
-"#;
-
-        let mut registry = SpecRegistry::new();
-        let (csdl, xml) = minimal_csdl();
-        registry.register_tenant("alpha", csdl, xml, &[("Order", GUARDED_IOA)]);
-        let tenant = TenantId::new("alpha");
-        let original_source = registry
-            .get_spec(&tenant, "Order")
-            .unwrap()
-            .ioa_source
-            .clone();
-
-        let (replacement_csdl, replacement_xml) = minimal_csdl();
-        let error = registry
-            .try_register_tenant(
-                "alpha",
-                replacement_csdl,
-                replacement_xml,
-                &[("Order", REACHABLE_IOA)],
-            )
-            .expect_err("changed reachable-state semantics require durable-state migration");
-
-        assert!(matches!(
-            error,
-            RegistryError::RuntimeInvariantMigrationRequired { .. }
-        ));
-        assert_eq!(
-            registry.get_spec(&tenant, "Order").unwrap().ioa_source,
-            original_source
-        );
-    }
-
-    #[test]
-    fn hot_swap_rejects_action_identity_collision_before_mutation() {
-        const UNIQUE_ACTIONS_IOA: &str = r#"
-[automaton]
-name = "Order"
-states = ["Unchecked", "Checked"]
-initial = "Unchecked"
-allow_indefinite_states = ["Unchecked", "Checked"]
-
-[[state]]
-name = "safe"
-type = "bool"
-initial = "true"
-
-[[action]]
-name = "CorruptUnchecked"
-kind = "input"
-from = ["Unchecked"]
-to = "Unchecked"
-effect = [{ type = "set_bool", var = "safe", value = "false" }]
-
-[[action]]
-name = "EnterChecked"
-kind = "input"
-from = ["Unchecked"]
-to = "Checked"
-effect = [{ type = "set_bool", var = "safe", value = "true" }]
-
-[[invariant]]
-name = "CheckedIsSafe"
-when = ["Checked"]
-assert = "safe"
-"#;
-
-        let mut registry = SpecRegistry::new();
-        let (csdl, xml) = minimal_csdl();
-        registry.register_tenant("alpha", csdl, xml, &[("Order", UNIQUE_ACTIONS_IOA)]);
-        let tenant = TenantId::new("alpha");
-        let original_spec = registry.get_spec(&tenant, "Order").unwrap();
-        let original_lock = original_spec.swap_controller().current();
-        let original_table = serde_json::to_value(&*original_spec.table()).unwrap();
-        let duplicate_actions = UNIQUE_ACTIONS_IOA.replacen(
-            "name = \"EnterChecked\"",
-            "name = \"CorruptUnchecked\"",
-            1,
-        );
-
-        let (replacement_csdl, replacement_xml) = minimal_csdl();
-        let error = registry
-            .try_register_tenant(
-                "alpha",
-                replacement_csdl,
-                replacement_xml,
-                &[("Order", &duplicate_actions)],
-            )
-            .expect_err("action identity changes L1 effect resolution");
-
-        assert!(matches!(
-            error,
-            RegistryError::RuntimeInvariantMigrationRequired { .. }
-        ));
-        let current = registry.get_spec(&tenant, "Order").unwrap();
-        assert_eq!(current.ioa_source, UNIQUE_ACTIONS_IOA);
-        assert!(Arc::ptr_eq(
-            &current.swap_controller().current(),
-            &original_lock
-        ));
-        assert_eq!(
-            serde_json::to_value(&*current.table()).unwrap(),
-            original_table
         );
     }
 }
